# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API for managing spy cats, missions, and agents",
        "title": "Spy Cat Agency API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/tokens/authentication/agents": {
            "post": {
                "description": "Authenticate an agent and return a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Create agent authentication token",
                "parameters": [
                    {
                        "description": "Agent Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.AuthenticationRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.TokenResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/main.ValidationErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/tokens/authentication/spy-cats": {
            "post": {
                "description": "Authenticate a spy cat and return a JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Create spy cat authentication token",
                "parameters": [
                    {
                        "description": "Spy Cat Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.AuthenticationRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.TokenResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/main.ValidationErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponseDoc"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.AuthenticationRequestDoc": {
            "description": "Request body for user authentication",
            "type": "object",
            "properties": {
                "name": {
                    "description": "User name\nExample: Agent Smith",
                    "type": "string"
                },
                "password": {
                    "description": "User password\nExample: password123",
                    "type": "string"
                }
            }
        },
...
        "main.ErrorResponseDoc": {
            "description": "Standard error response format",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message\nExample: error message",
                    "type": "string"
                }
            }
        },
...
        "main.TokenDoc": {
            "description": "Authentication token entity",
            "type": "object",
            "properties": {
                "expiry": {
                    "description": "Token expiry date\nExample: 2024-01-01T00:00:00Z",
                    "type": "string"
                },
                "plaintext": {
                    "description": "Token plaintext value\nExample: ABCDEF123456",
                    "type": "string"
                },
                "scope": {
                    "description": "Token scope\nExample: authentication",
                    "type": "string"
                },
                "user_id": {
                    "description": "User ID associated with the token\nExample: 1",
                    "type": "integer"
                }
            }
        },
        "main.TokenResponseDoc": {
            "description": "Response containing an authentication token",
            "type": "object",
            "properties": {
                "authentication_token": {
                    "description": "Authentication token data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/main.TokenDoc"
                        }
                    ]
                }
            }
        },
...
        "main.ValidationErrorResponseDoc": {
            "description": "Validation error response format with field-specific errors",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Map of field names to error messages\nExample: {\"name\": \"must be provided\", \"breed\": \"invalid breed\"}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}